requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
torch==2.1.0
transformers==4.35.0
scikit-learn==1.3.0
sentence-transformers==2.2.2
pandas==2.1.3
numpy==1.24.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.6
pyyaml==6.0.1
elasticsearch==8.11.0
prometheus-client==0.19.0
chardet==5.2.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
joblib==1.3.2
seaborn==0.12.2
matplotlib==3.7.2
